split tables, moved to sis/
